fix type overflow

+++ src/blsct/range_proof/bulletproofs/range_proof_logic.cpp
@@ -370,7 +370,7 @@ AmountRecoveryResult<T> RangeProofLogic<T>::RecoverAmounts(
     // will contain result of successful requests only
     std::vector<range_proof::RecoveredData<T>> xs;
 
-    size_t i = -1;
+    int i = -1;
 
     for (const AmountRecoveryRequest<T>& req : reqs) {
         const range_proof::Generators<T> gens = m_common.Gf().GetInstance(req.token_id);
@@ -423,7 +423,7 @@ AmountRecoveryResult<T> RangeProofLogic<T>::RecoverAmounts(
         auto msg_amt = maybe_msg_amt.value();
 
         auto x = range_proof::RecoveredData<T>(
-            i,
+            (size_t)i,
             msg_amt.amount,
             req.nonce.GetHashWithSalt(100), // gamma for vs[0]
             msg_amt.msg);
